addressing @gregcaporaso comments
